find products

+++ server/database/migration/01_create_table_products.up.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS tb_products (
   name TEXT UNIQUE NOT NULL,
   brand VARCHAR(50) UNIQUE NOT NULL,
   description TEXT NOT NULL,
-  price NUMERIC(10, 2) NOT NULL,
+  price REAL NOT NULL,
   stock SMALLINT NOT NULL,
   created_at TIMESTAMP NOT NULL
 );
